fix running tests
